Refactor user routes
